refactor: break out utils
