+++ lib/rules/utils/ast-utils.js
@@ -9,7 +9,7 @@
 // Typedefs
 //------------------------------------------------------------------------------
 
-/** @typedef {import('../../shared/types').ExclusifyUnion<import('estree').Node>} ASTNode */
+/** @typedef {import('../../shared/types').ASTNode>} ASTNode */
 /** @typedef {import('estree').Comment & { type: 'Block'}} Block */
 /** @typedef {import('estree').Comment & { type: 'Line'}} Line */
 /** @typedef {import('eslint').AST.SourceLocation} SourceLocation */
